Moved "tk::spline" and "asciiplotter" into seperate folders.
