+++ src/utility/asciiplotter/README.md
@@ -1,0 +1,6 @@
+**joehood::asciiplotter** is a C++ package used to construct ASCII plots in text format as a quick way of checking the results. 
+
+The source code can be found in [joehood's GitHub](https://github.com/joehood/asciiplotter).
+
+#### Modifications:
+- Converted some ```int``` to ```std::size_t``` to avoid some warnings.
