+++ src/utility/asciiplotter/asciiplotter.cpp
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include <BSMPT/utility/asciiplotter.h>
+#include <BSMPT/utility/asciiplotter/asciiplotter.h>
 
 using namespace std;
