+++ src/utility/spline/README.md
@@ -1,0 +1,6 @@
+**tk::spline** is a C++ package used to construct cubic splines. 
+
+The source code can be found in [ttk592's GitHub](https://github.com/ttk592/spline) and the documentation at https://kluge.in-chemnitz.de/.
+
+#### Modifications:
+- Split the code into declarations and implementations.
